used props, useEffect hook, and useState hook to debug
