--- src/About.tsx
@@ -1,44 +1,0 @@
-const About = () => {
-    return (
-        <>
-          <div className="about-content-flex">
-                    <section className="about-section2-bg">
-                        <div className="about-section-content">
-                            <h1>About Me</h1>
-                            <p>I've studied Multimedia and Music since 2018. My journey started very early when I attended many coding and game design camps in my youth. One notable program was ID Tech that hosted a game design program at Georgia Tech. This was my first time interacting with software such as Maya and this was where I found my passion for creativity and digital design. The name Austral Lunaris was later created in my love for the moon and the stars and base many of my creations off of it.</p>
-                            <p>I started off making a music promotion channel on YouTube that was suggested by a friend. This lead to me learning the basics of digital aspects such as picture dimensions for video editing. This eventually led to me learning animations and edits in order to make my my videos stand out.</p>
-                            <p>My goal is to make what I'm doing into a full time career and to give back to the art community. It's always satisfying to see how excited artists become when they are able to see their creations come to life.</p>
-                            <p>All Links: https://linktr.ee/Austral.Lunaris</p>
-                            <div className="line-seporator"></div>
-                            <div className="skills-text">Skills</div>
-                            <div className="skills">
-                                <p className="skills-grid1">Video Editing</p>
-                                <p className="skills-grid2">Graphic Design</p>
-                                <p className="skills-grid3">Animations</p>
-                                <p className="skills-grid4">HTML & CSS</p>
-                                <p className="skills-grid5">Typography</p>
-                                <p className="skills-grid6">Music Production</p>
-                                <p className="skills-grid7">Java Script</p>
-                            </div>
-                            <div className="line-seporator"></div>
-                            <div className="software-text">Software</div>
-                            <div className="software">
-                                <p>Adobe Photoshop</p>
-                                <p>Adobe After-Effects</p>
-                                <p>Adobe Illustrator</p>
-                                <p>Adobe In-Design</p>
-                                <p>Spine-2D</p>
-                                <p>FL-Studio</p>
-                                <p>Maya-Autodesk</p>
-                                <p>Davinci Resolve</p>
-                                <p>Visual Studio Code</p>
-                            </div>
-                        </div>
-                    </section>
-            </div>     
-
-        </>
-    )
-}
-
-export default About;
