--- src/Animations.tsx
@@ -1,27 +1,0 @@
-const Animations = () => {
-    return (
-        <>
-        <div className="gallery-section">
-        <h1>Animations</h1>
-            <div className="gallery-container">
-                <div className="animations-video">
-                    <video src="./Videos/World_is_Yours_10_Seconds.mp4" controls muted> </video>
-                    <video src="./Videos/Ageturnerx_Anna_Choi_Cloudscape_1_1.mp4" controls muted> </video>
-                    <video src="./Videos/Naradala_There_is_a_Dragon_up There_1.mp4" controls muted> </video>
-                    <video src="./Videos/Narardala_A Bird That Does not Sing_1.mp4" controls muted> </video>
-                    <video src="./Videos/JESSAUDREY_Trancend_1.mp4" controls muted> </video>
-                    <video src="./Videos/YUMEFOREVER_Never_Cry_Again_Single.mp4" controls muted> </video>
-                    <video src="./Videos/Rey_HeavenSward Funk_1.mp4" controls muted> </video>
-                    <video src="./Videos/17929384093574157.mp4" controls muted> </video>
-                    <video src="./Videos/17878932143344437.mp4" controls muted> </video>
-                    <video src="./Videos/Schiller Productions Logo Intro Ver.2.mp4" controls muted> </video>
-                    <video src="./Videos/Maroon_Fest_Logo_Main_1.mp4" controls muted> </video>
-                    <video src="./Videos/Logo.mp4" controls muted> </video>
-                </div>    
-            </div>   
-        </div>
-        </>
-    )
-}
-
-export default Animations;
