--- src/Designs.tsx
@@ -1,17 +1,0 @@
-const Designs = () => {
-    return(
-        <>
-        <div className="gallery-section">
-            <h1>Designs</h1>
-            <div className="gallery-container">
-                <div className="gallery-picture">
-                    <img src="./pictures/Chalks_BBQ_Banner.png"></img>
-                    <img src="./pictures/Chalks_BBQ.png"></img>
-                </div>
-            </div>
-        </div>
-        </>
-    )
-}
-
-export default Designs;
